+++ src/extensions/taction/tactionext.cpp
@@ -31,6 +31,7 @@
 #include "tactionext.h"
 
 #include "debughandler.h"
+#include "building.h"
 #include "house.h"
 #include "houseext.h"
 #include "housetype.h"
@@ -44,6 +45,7 @@
 #include "tag.h"
 #include "tagtype.h"
 #include "techno.h"
+#include "technoext.h"
 #include "tibsun_inline.h"
 #include "trigger.h"
 #include "triggertype.h"
@@ -56,7 +58,7 @@ TActionClass::ActionDescriptionStruct TActionClassExtension::ExtActionDescriptio
     { "Give Credits", "Gives or removes credits from the specified house. A positive amount gives money, a negative amount subtracts it." },
     { "Enable Short Game", "Enables Short Game. Players will lose if all buildings are destroyed." },
     { "Disable Short Game", "Disables Short Game. Players can continue playing even after all buildings are destroyed." },
-    { "Unused Action", "This action does nothing. Originally used to display the difficulty in ts-patches." },
+    { "Create Building At", "Places a building at given waypoint position." },
     { "Destroy all of...", "Kills everything of the specified house and marks them as defeated." },
     { "Make Elite", "All technos attached to this trigger will be promoted to elite status." },
     { "Enable Ally Reveal", "Enables Ally Reveal, allowing allied players to see each other's explored areas." },
@@ -83,6 +85,7 @@ TActionClass::ActionDescriptionStruct TActionClassExtension::ExtActionDescriptio
     { "Enable templated text", "Displays a line of text on the screen with variable substitution. The text may include placeholders like {{g_variableName}} or {{l_variableName}}, which are replaced with the corresponding global or local variable values. Color `-1` uses the color of the player's house." },
     { "Disable templated text", "Removes the currently active templated text from the screen." },
     { "Adjust House Modifier", "Adjusts a house modifier by given percentage points." },
+    { "Apply Iron Curtain", "Applies Iron Curtain to attached objects. Can optionally bypass legality checks." },
 };
 
 
@@ -301,6 +304,7 @@ bool TActionClassExtension::Execute(HouseClass* house, ObjectClass* object, Trig
         EXT_DISPATCH(GIVE_CREDITS);
         EXT_DISPATCH(ENABLE_SHORT_GAME);
         EXT_DISPATCH(DISABLE_SHORT_GAME);
+        EXT_DISPATCH(CREATE_BUILDING_AT);
         EXT_DISPATCH(HOUSE_DESTROY_ALL);
         EXT_DISPATCH(MAKE_ELITE);
         EXT_DISPATCH(ENABLE_ALLYREVEAL);
@@ -327,13 +331,7 @@ bool TActionClassExtension::Execute(HouseClass* house, ObjectClass* object, Trig
         EXT_DISPATCH(ENABLE_TEMPLATED_TEXT);
         EXT_DISPATCH(DISABLE_TEMPLATED_TEXT);
         EXT_DISPATCH(ADJUST_HOUSE_MODIFIER);
-
-        /**
-         *  Used to print the current difficulty in ts-patches, available to be repurposed.
-         */
-    case EXT_TACTION_UNUSED1:
-        success = true;
-        break;
+        EXT_DISPATCH(APPLY_IRON_CURTAIN);
 
         /**
          *  Unexpected TActionType.
@@ -704,6 +702,52 @@ bool TActionClassExtension::Do_DISABLE_SHORT_GAME(HouseClass* house, ObjectClass
     Session.Options.ShortGame = false;
 
     return true;
+}
+
+
+/**
+ *  Places a building at given waypoint position.
+ *
+ *  @author: Rampastring
+ */
+bool TActionClassExtension::Do_CREATE_BUILDING_AT(HouseClass* house, ObjectClass* object, TriggerClass* trig, const Cell& cell)
+{
+    Cell wpcell = ScenExtension->Waypoint_Cell(This()->EffectLocation);
+
+    if (wpcell != CELL_NONE) {
+        HouseClass* hptr = HouseClassExtension::House_From_HousesType(This()->Data.House);
+
+        int buildingtypeid = This()->TriggerRect.X;
+        bool forced = This()->TriggerRect.Y > 0;
+        BuildingTypeClass* btc = BuildingTypes[buildingtypeid];
+
+        bool success = false;
+
+        if (forced) {
+            ScenarioInit++;
+            success = btc->Create_And_Place(wpcell, hptr);
+            ScenarioInit--;
+        } else {
+            // Create_And_Place does not play buildup anim
+            BuildingClass* building = new BuildingClass(btc, hptr);
+
+            if (building != nullptr) {
+                building->Assign_Mission(MISSION_CONSTRUCTION);
+                success = building->Unlimbo(wpcell.As_Coord());
+
+                if (!success) {
+                    delete building;
+                } else {
+                    building->Revealed(hptr);
+                    building->IsReadyToCommence = true;
+                }
+            }
+        }
+
+        return success;
+    }
+
+    return false;
 }
 
 
@@ -1534,3 +1578,39 @@ bool TActionClassExtension::Do_ADJUST_HOUSE_MODIFIER(HouseClass* house, ObjectCl
     return true;
 }
 
+
+/**
+ *  Applies the Iron Curtain to attached objects.
+ *
+ *  @author: Rampastring
+ */
+bool TActionClassExtension::Do_APPLY_IRON_CURTAIN(HouseClass* house, ObjectClass* object, TriggerClass* trig, const Cell& cell)
+{
+    HouseClassExtension* houseext = Extension::Fetch(house);
+
+    // Check for legality, unless this is forced.
+    bool forced = This()->Data.Bool;
+    if (!forced) {
+        if (!houseext->Can_Use_Iron_Curtain()) {
+            // If the application is not forced and the house is unable to use the Iron Curtain, skip.
+            return true;
+        }
+    }
+
+    /**
+     *  Iterate all technos, and if their tag is attached to this trigger, apply Iron Curtain on them.
+     */
+    for (int i = 0; i < Technos.Count(); i++) {
+        TechnoClass* techno = Technos[i];
+
+        if (techno->IsActive && techno->IsDown && !techno->IsInLimbo) {
+            if (techno->Tag && techno->Tag->Is_Trigger_Attached(trig)) {
+                TechnoClassExtension* technoext = Extension::Fetch(techno);
+                technoext->Iron_Curtain_Me(true);
+            }
+        }
+    }
+
+    houseext->Expend_Iron_Curtain();
+    return true;
+}