+++ src/extensions/techno/technoext.h
@@ -74,6 +74,8 @@ class TechnoClassExtension : public RadioClassExtension
         bool Can_Opportunity_Fire() const;
         bool Opportunity_Fire();
 
+        bool Iron_Curtain_Me(bool forced);
+
     private:
         const TechnoTypeClass *Techno_Type_Class() const;
         const TechnoTypeClassExtension *Techno_Type_Class_Ext() const;
@@ -129,4 +131,9 @@ class TechnoClassExtension : public RadioClassExtension
          *  The idle wake animation attached to this object.
          */
         AnimClass* IdleWakeAnim;
+
+        /**
+         *  The countdown until the object's Iron Curtain effect fades away.
+         */
+        CDTimerClass<FrameTimerClass> IronCurtainTimer;
 };